% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.signer_operations.R
\name{get_signing_platform}
\alias{get_signing_platform}
\title{Returns information on a specific signing platform}
\usage{
get_signing_platform(
  platformId = "string"
)
}
\arguments{
\item{platformId}{[required] The ID of the target signing platform.}
}
\description{
Returns information on a specific signing platform.
}
\examples{

}
