% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.signer_operations.R
\name{get_signing_profile}
\alias{get_signing_profile}
\title{Returns information on a specific signing profile}
\usage{
get_signing_profile(profileName)
}
\arguments{
\item{profileName}{[required] The name of the target signing profile.}
}
\description{
Returns information on a specific signing profile.
}
\section{Accepted Parameters}{
\preformatted{get_signing_profile(
  profileName = "string"
)
}
}

