% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.signer_operations.R
\name{list_signing_jobs}
\alias{list_signing_jobs}
\title{Lists all your signing jobs}
\usage{
list_signing_jobs(
  status = "InProgress"|"Failed"|"Succeeded",
  platformId = "string",
  requestedBy = "string",
  maxResults = 123,
  nextToken = "string"
)
}
\arguments{
\item{status}{A status value with which to filter your results.}

\item{platformId}{The ID of microcontroller platform that you specified for the distribution of your code image.}

\item{requestedBy}{The IAM principal that requested the signing job.}

\item{maxResults}{Specifies the maximum number of items to return in the response. Use this parameter when paginating results. If additional items exist beyond the number you specify, the \code{nextToken} element is set in the response. Use the \code{nextToken} value in a subsequent request to retrieve additional items.}

\item{nextToken}{String for specifying the next set of paginated results to return. After you receive a response with truncated results, use this parameter in a subsequent request. Set it to the value of \code{nextToken} from the response that you just received.}
}
\description{
Lists all your signing jobs. You can use the \code{maxResults} parameter to limit the number of signing jobs that are returned in the response. If additional jobs remain to be listed, AWS Signer returns a \code{nextToken} value. Use this value in subsequent calls to \code{ListSigningJobs} to fetch the remaining values. You can continue calling \code{ListSigningJobs} with your \code{maxResults} parameter and with new values that AWS Signer returns in the \code{nextToken} parameter until all of your signing jobs have been returned.
}
\examples{

}
