% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.signer_operations.R
\name{list_signing_profiles}
\alias{list_signing_profiles}
\title{Lists all available signing profiles in your AWS account}
\usage{
list_signing_profiles(includeCanceled = NULL, maxResults = NULL,
  nextToken = NULL)
}
\arguments{
\item{includeCanceled}{Designates whether to include profiles with the status of \code{CANCELED}.}

\item{maxResults}{The maximum number of profiles to be returned.}

\item{nextToken}{Value for specifying the next set of paginated results to return. After you receive a response with truncated results, use this parameter in a subsequent request. Set it to the value of \code{nextToken} from the response that you just received.}
}
\description{
Lists all available signing profiles in your AWS account. Returns only profiles with an \code{ACTIVE} status unless the \code{includeCanceled} request field is set to \code{true}. If additional jobs remain to be listed, AWS Signer returns a \code{nextToken} value. Use this value in subsequent calls to \code{ListSigningJobs} to fetch the remaining values. You can continue calling \code{ListSigningJobs} with your \code{maxResults} parameter and with new values that AWS Signer returns in the \code{nextToken} parameter until all of your signing jobs have been returned.
}
\section{Accepted Parameters}{
\preformatted{list_signing_profiles(
  includeCanceled = TRUE|FALSE,
  maxResults = 123,
  nextToken = "string"
)
}
}

